Sesion1.3 prac0

+++ src/main/java/ecuacion/EcuacionLineal.java
@@ -1,15 +1,15 @@
 package ecuacion;
 
 public class EcuacionLineal {
-    private int x;
-    private int indep;
+    private double x;
+    private double indep;
 
     public EcuacionLineal(int x, int indep) {
         this.x = x;
         this.indep = indep;
     }
 
-    public int resolverEcuacion () throws ArithmeticException {
+    public double resolverEcuacion () throws ArithmeticException {
         if (this.x != 0) { return (-this.indep/this.x); }
         else throw new ArithmeticException();
     }
